wake-lock/wakelock-api.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Wake Lock API Test</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="author" title="Wanming Lin" href="mailto:wanming.lin@intel.com">
<link rel="help" href="https://w3c.github.io/wake-lock/">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>

<script>

promise_test(async t => {
  assert_true("getWakeLock" in navigator, "Wake Lock API is present");
  const wakeLock = await navigator.getWakeLock("screen");
  assert_true(wakeLock instanceof WakeLock, "wakeLock is a WakeLock");
  assert_true(wakeLock.type instanceof WakeLockType, "wakeLock.type is a WakeLockType");
  assert_equals(typeof wakeLock.active, "boolean", "the type of wakeLock.active is boolean");
}, "Test that the Wake Lock API is correct");

</script>
